refactor(core): hoist shared import-stub range helpers into ArchBackend
X86Backend and AArch64Backend each carried a byte-for-byte identical
_getPltRanges/_getMachoStubRanges/_getImportStubRanges trio for resolving
ELF PLT and Mach-O stub address ranges. Moves them onto the shared
ArchBackend base class so both backends inherit one implementation instead
of drifting independently; each backend keeps only its architecture-specific
slot-decode logic (_resolveImportSlot / _resolvePltGotSlot).

Validation: make lint, make test
